[symint] Fix SizesAndStridesTest to not use negative sizes/strides
With SymInt we are using the negative space of `int64_t` in our internal
representation. `SizesAndStridesTest` breaks this because it initializes
`SizesAndStrides` with negative sizes/strides. This PR fixes that.

As an aside: feels like `SizesAndStrides` (and `SymInt`) should really
take a uint64_t, but that would be BC-breaking so I don't do it here.

ghstack-source-id: eeb06ddbde4b89f31a6ae7cb94ca5e25286dfae4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77820